Improve string escaping in template-literals
I didn't like how escape sequences would be converted to the actual
characters after running this transform, so I looked into how that is
represented in string literals and noticed that we have access to a
`raw` property that contains the escaped string that we need, with some
additional quotation marks around it. After stripping off these
quotation marks and escaping ${ to prevent new interpolation, we have
much better string escaping.